Organize schemas into correct places
