+++ kasa/tests/fakeprotocol_iot.py
@@ -1,176 +1,11 @@
 import copy
 import logging
-import re
-
-from voluptuous import (
-    REMOVE_EXTRA,
-    All,
-    Any,
-    Coerce,  # type: ignore
-    Invalid,
-    Optional,
-    Range,
-    Schema,
-)
 
 from ..deviceconfig import DeviceConfig
 from ..iotprotocol import IotProtocol
 from ..xortransport import XorTransport
 
 _LOGGER = logging.getLogger(__name__)
-
-
-def check_int_bool(x):
-    if x != 0 and x != 1:
-        raise Invalid(x)
-    return x
-
-
-def check_mac(x):
-    if re.match("[0-9a-f]{2}([-:])[0-9a-f]{2}(\\1[0-9a-f]{2}){4}$", x.lower()):
-        return x
-    raise Invalid(x)
-
-
-def check_mode(x):
-    if x in ["schedule", "none", "count_down"]:
-        return x
-
-    raise Invalid(f"invalid mode {x}")
-
-
-def lb_dev_state(x):
-    if x in ["normal"]:
-        return x
-
-    raise Invalid(f"Invalid dev_state {x}")
-
-
-TZ_SCHEMA = Schema(
-    {"zone_str": str, "dst_offset": int, "index": All(int, Range(min=0)), "tz_str": str}
-)
-
-CURRENT_CONSUMPTION_SCHEMA = Schema(
-    Any(
-        {
-            "voltage": Any(All(float, Range(min=0, max=300)), None),
-            "power": Any(Coerce(float, Range(min=0)), None),
-            "total": Any(Coerce(float, Range(min=0)), None),
-            "current": Any(All(float, Range(min=0)), None),
-            "voltage_mv": Any(
-                All(float, Range(min=0, max=300000)), int, None
-            ),  # TODO can this be int?
-            "power_mw": Any(Coerce(float, Range(min=0)), None),
-            "total_wh": Any(Coerce(float, Range(min=0)), None),
-            "current_ma": Any(
-                All(float, Range(min=0)), int, None
-            ),  # TODO can this be int?
-            "slot_id": Any(Coerce(int, Range(min=0)), None),
-        },
-        None,
-    )
-)
-
-# these schemas should go to the mainlib as
-# they can be useful when adding support for new features/devices
-# as well as to check that faked devices are operating properly.
-PLUG_SCHEMA = Schema(
-    {
-        "active_mode": check_mode,
-        "alias": str,
-        "dev_name": str,
-        "deviceId": str,
-        "feature": str,
-        "fwId": str,
-        "hwId": str,
-        "hw_ver": str,
-        "icon_hash": str,
-        "led_off": check_int_bool,
-        "latitude": Any(All(float, Range(min=-90, max=90)), 0, None),
-        "latitude_i": Any(
-            All(int, Range(min=-900000, max=900000)),
-            All(float, Range(min=-900000, max=900000)),
-            0,
-            None,
-        ),
-        "longitude": Any(All(float, Range(min=-180, max=180)), 0, None),
-        "longitude_i": Any(
-            All(int, Range(min=-18000000, max=18000000)),
-            All(float, Range(min=-18000000, max=18000000)),
-            0,
-            None,
-        ),
-        "mac": check_mac,
-        "model": str,
-        "oemId": str,
-        "on_time": int,
-        "relay_state": int,
-        "rssi": Any(int, None),  # rssi can also be positive, see #54
-        "sw_ver": str,
-        "type": str,
-        "mic_type": str,
-        "updating": check_int_bool,
-        # these are available on hs220
-        "brightness": int,
-        "preferred_state": [
-            {"brightness": All(int, Range(min=0, max=100)), "index": int}
-        ],
-        "next_action": {"type": int},
-        "child_num": Optional(Any(None, int)),  # TODO fix hs300 checks
-        "children": Optional(list),  # TODO fix hs300
-        # TODO some tplink simulator entries contain invalid (mic_mac, _i variants for lat/lon)
-        # Therefore we add REMOVE_EXTRA..
-        # "INVALIDmac": Optional,
-        # "INVALIDlatitude": Optional,
-        # "INVALIDlongitude": Optional,
-    },
-    extra=REMOVE_EXTRA,
-)
-
-LIGHT_STATE_SCHEMA = Schema(
-    {
-        "brightness": All(int, Range(min=0, max=100)),
-        "color_temp": int,
-        "hue": All(int, Range(min=0, max=360)),
-        "mode": str,
-        "on_off": check_int_bool,
-        "saturation": All(int, Range(min=0, max=100)),
-        "dft_on_state": Optional(
-            {
-                "brightness": All(int, Range(min=0, max=100)),
-                "color_temp": All(int, Range(min=0, max=9000)),
-                "hue": All(int, Range(min=0, max=360)),
-                "mode": str,
-                "saturation": All(int, Range(min=0, max=100)),
-            }
-        ),
-        "err_code": int,
-    }
-)
-
-BULB_SCHEMA = PLUG_SCHEMA.extend(
-    {
-        "ctrl_protocols": Optional(dict),
-        "description": Optional(str),  # TODO: LBxxx similar to dev_name
-        "dev_state": lb_dev_state,
-        "disco_ver": str,
-        "heapsize": int,
-        "is_color": check_int_bool,
-        "is_dimmable": check_int_bool,
-        "is_factory": bool,
-        "is_variable_color_temp": check_int_bool,
-        "light_state": LIGHT_STATE_SCHEMA,
-        "preferred_state": [
-            {
-                "brightness": All(int, Range(min=0, max=100)),
-                "color_temp": int,
-                "hue": All(int, Range(min=0, max=360)),
-                "index": int,
-                "saturation": All(int, Range(min=0, max=100)),
-            }
-        ],
-    }
-)
 
 
 def get_realtime(obj, x, *args):
